+++ lognflow/logviewer.py
@@ -1,5 +1,6 @@
 import pathlib
 import numpy as np
+from .utils import replace_all
 
 class logviewer:
     """ log viewer
@@ -19,7 +20,8 @@ class logviewer:
         assert self.log_dir.is_dir(), \
             f'lognflow.logviewer| No such directory: '+ str(self.log_dir)
         self.logger = logger
-        
+        self.load = self.get_single
+    
     def get_text(self, log_name='main_log'):
         """ get text log files
             Given the log_name, this function returns the text therein.
@@ -41,25 +43,18 @@ class logviewer:
                 txt = txt[0]
             return txt
 
-    def get_single(self, var_name, single_shot_index = -1, 
-                     suffix = None):
-        """ get a single variable
-            return the value of a saved variable.
+    def get_flist(self, var_name, suffix = None):
+        """ get list of files
+            return the list of files for a saved variable.
 
             Parameters
             ----------
             :param var_name:
                 variable name
-            :param single_shot_index:
-                If there are many snapshots of a variable, this input can
-                limit the returned to a set of indices.
             :param suffix:
                 If there are different suffixes availble for a variable
                 this input needs to be set. npy, npz, mat, and torch are
                 supported.
-                
-            .. note::
-                when reading a MATLAB file, the output is a dictionary.
         """
         var_name = var_name.replace('\t', '\\t').replace('\n', '\\n')\
             .replace('\r', '\\r').replace('\b', '\\b')
@@ -77,8 +72,7 @@ class logviewer:
                 suffix = '.np*'
 
         suffix = suffix.strip('.')        
-        assert single_shot_index == int(single_shot_index), \
-                    f'single_shot_index {single_shot_index} must be an integer'
+
         flist = []            
         if((self.log_dir / var_name).is_file()):
             flist = [self.log_dir / var_name]
@@ -88,15 +82,18 @@ class logviewer:
             _var_name = (self.log_dir / var_name).name
             _var_dir = (self.log_dir / var_name).parent
             search_patt = f'{_var_name}*.{suffix}'
-            while('**' in search_patt):
-                search_patt = search_patt.replace('**', '*')
+            search_patt = replace_all(search_patt, '**', '*')
             flist = list(_var_dir.glob(search_patt))
             if(len(flist) == 0):
-                flist = list(_var_dir.glob(f'{_var_name}*.*'))
+                search_patt = f'{_var_name}*.*'
+                search_patt = replace_all(search_patt, '**', '*')
+                flist = list(_var_dir.glob(search_patt))
                 if(len(flist) > 0):
-                    self.logger('Can not find the file with the given suffix, '\
-                                +'but found some with a different suffix, '\
-                                +f'one file is: {flist[single_shot_index]}')
+                    self.logger(
+                        'I Can not find the file with the given suffix, '\
+                        + 'but found some with a different suffix, '\
+                        + f'one file is: {flist[-1]}. This is what I'\
+                        + ' will return.' )
                     
         if(len(flist) > 0):
             flist.sort()
@@ -106,52 +103,80 @@ class logviewer:
                 flist = list(var_dir.glob('*.*'))
             if(len(flist) > 0):
                 flist.sort()
-            else:
-                self.logger('No such variable')
-                return
-        var_path = flist[single_shot_index]
+        return flist
+            
+    def get_single(self, var_name, file_index = -1, suffix = None):
+        """ get a single variable
+            return the value of a saved variable.
+
+            Parameters
+            ----------
+            :param var_name:
+                variable name
+            :param file_index:
+                If there are many snapshots of a variable, this input can
+                limit the returned to a set of indices.
+            :param suffix:
+                If there are different suffixes availble for a variable
+                this input needs to be set. npy, npz, mat, and torch are
+                supported.
                 
-        if(var_path.is_file()):
-            self.logger(f'Loading {var_path}')
-            if(var_path.suffix == '.npz'):
-                buf = np.load(var_path)
+            .. note::
+                when reading a MATLAB file, the output is a dictionary.
+        """
+        assert file_index == int(file_index), \
+                    f'file_index {file_index} must be an integer'
+        flist = self.get_flist(var_name, suffix)
+        var_path = None
+        if flist:
+            var_path = flist[file_index]
+    
+            if(var_path.is_file()):
+                self.logger(f'Loading {var_path}')
+                if(var_path.suffix == '.npz'):
+                    buf = np.load(var_path)
+                    try:
+                        time_array = buf['time_array']
+                        n_logs = (time_array > 0).sum()
+                        time_array = time_array[:n_logs]
+                        data_array = buf['data_array']
+                        data_array = data_array[:n_logs]
+                        return((time_array, data_array))
+                    except:
+                        return(buf)
+                if(var_path.suffix == '.npy'):
+                    return(np.load(var_path))
+                if(var_path.suffix == '.mat'):
+                    from scipy.io import loadmat
+                    return(loadmat(var_path))
+                if( (var_path.suffix == '.txt')
+                   |(var_path.suffix == '.pdb')
+                   |(var_path.suffix == '.json')):
+                    with open(var_path) as f_txt:
+                        return(f_txt.read())
+                if((var_path.suffix == '.tif') | (var_path.suffix == '.tiff')):
+                    from tifffile import imread
+                    return(imread(var_path))
+                if(var_path.suffix == '.torch'):      
+                    from torch import load as torch_load 
+                    return(torch_load(var_path))
                 try:
-                    time_array = buf['time_array']
-                    n_logs = (time_array > 0).sum()
-                    time_array = time_array[:n_logs]
-                    data_array = buf['data_array']
-                    data_array = data_array[:n_logs]
-                    return((time_array, data_array))
+                    from matplotlib.pyplot import imread
+                    img = imread(var_path)
+                    return(img)
                 except:
-                    return(buf)
-            if(var_path.suffix == '.npy'):
-                return(np.load(var_path))
-            if(var_path.suffix == '.mat'):
-                from scipy.io import loadmat
-                return(loadmat(var_path))
-            if(var_path.suffix == '.txt'):
-                with open(var_path) as f_txt:
-                    return(f_txt.read())
-            if((var_path.suffix == '.tif') | (var_path.suffix == '.tiff')):
-                from tifffile import imread
-                return(imread(var_path))
-            if(var_path.suffix == '.torch'):      
-                from torch import load as torch_load 
-                return(torch_load(var_path))
-            try:
-                from matplotlib.pyplot import imread
-                img = imread(var_path)
-                return(img)
-            except:
-                pass
-        else:
-            self.logger(f'{var_name} not found.')
-            return
-   
-    def get_stack_of_files(self,
+                    var_path = None
+            else:
+                var_path = None
+                
+        if var_path is None:
+            self.logger(f'{var_name} does not resemble the name of any file '
+                        f'or directory in the log_dir: {self.log_dir}')
+    
+    def get_stack_of_files(self, 
         var_name = None, flist = [], suffix = None,
         return_data = False, return_flist = True, read_func = None,
-        data_makes_a_block = False, verbose = True):
+        data_makes_a_block = True, verbose = True):
        
         """ Get list or data of all files in a directory
        
@@ -194,48 +219,15 @@ class logviewer:
                 flist is type pathlib.Path
            
         """
-        if suffix is None:
-            var_name_split = var_name.split('.')
-            if len(var_name_split) > 1:
-                suffix = var_name_split[-1]
-                name_before_suffix = var_name_split[:-1]
-                if((len(name_before_suffix) == 1) &
-                   (name_before_suffix[0] == '')):
-                    var_name = '*'
-                else:
-                    var_name = ('.').join(var_name_split[:-1])
-            else:
-                suffix = '*'
-
-        suffix = suffix.strip('.')
         if not flist:
-            assert var_name is not None, \
-                ' The file list is empty. Please provide the ' \
-                + 'variable name or a non-empty file list.'
-            var_dir = self.log_dir / var_name
-            if(var_dir.is_dir()):
-                var_fname = None
-                flist = list(var_dir.glob(f'*.{suffix}'))
-            else:
-                var_fname = var_dir.name
-                var_dir = var_dir.parent
-                patt = f'{var_fname}*.{suffix}'
-                while('**' in patt):
-                    patt = patt.replace('**', '*')
-                flist = list(var_dir.glob(patt))
-        else:
-            var_dir = flist[0].parent
-            assert var_dir.is_dir(),\
-                f'The directory {var_dir} for the '\
-                + 'provided list cannot be accessed.'
-           
+            flist = self.get_flist(var_name, suffix)
+        
         if flist:
             flist.sort()
             n_files = len(flist)
             if((not return_data) & return_flist):
                 return(flist)
            
-            ######### asked for data ########
             if(read_func is None):
                 try:
                     fdata = np.load(flist[0])
@@ -257,10 +249,20 @@ class logviewer:
                     dataset = np.zeros((n_files, ) + fdata.shape,
                                        dtype=fdata.dtype)
                     if(verbose):
-                        self.logger(f'logviewer: Reading dataset from {var_dir}'
+                        self.logger(f'logviewer: Reading dataset from '
+                                    f'{flist[0].parent}'
                                     f', the shape would be: {dataset.shape}')
                     for fcnt, fpath in enumerate(flist):
-                        dataset[fcnt] = read_func(fpath)
+                        try:
+                            dataset[fcnt] = read_func(fpath)
+                        except e:
+                            self.logger(
+                                'I cannot concatenate the data I just read on'
+                                ' top of last files data. You need to set'
+                                ' data_makes_a_block = False in the input'
+                                f' of get_stack_of_files({var_name}, ...)'
+                                ' This will return a list of data of files')
+                            raise e
                 else:
                     dataset = [read_func(fpath) for fpath in flist]
                 if(return_flist):
@@ -315,46 +317,3 @@ class logviewer:
 
     def __bool__(self):
         return self.log_dir.is_dir()
-
-def str2type(_element):
-    if _element[0] == '\'':
-        return _element[1:-1]
-    else:
-        try:
-            return int(_element)
-        except ValueError:
-            try:
-                return float(_element)
-            except ValueError:
-                pass
-    return _element
-
-def text_to_object(txt):
-    """ Read a list or dict that was sent to write to text e.g. via log_single:
-    As you may have tried, it is possible to send a Pythonic list to a text file
-    the list will be typed there with [ and ] and ' and ' for strings with ', '
-    in between. In this function we will merely return the actual content
-    of the original list.
-    Now if the type the element of the list was string, it would put ' and ' in
-    the text file. But if it is a number, no kind of punctuation or sign is 
-    used. by write(). We support int or float. Otherwise the written text
-    will be returned as string with any other wierd things attached to it.
-    
-    """
-    if(txt[0] == '['):
-        txt = txt.strip('[').strip(']')
-        txt = txt.split(', ')
-        obj_out = txt
-        for cnt, _element in enumerate(txt):
-            obj_out[cnt] = str2type(_element)
-    elif(txt[0] == '{'):
-        txt = txt.strip('{').strip('}')
-        txt = txt.split(', ')
-        obj_out = dict()
-        for cnt, _element in enumerate(txt):
-            _element_key = str2type(_element.split(': ')[0])
-            _element_value = str2type(_element.split(': ')[1])
-            obj_out[_element_key] = _element_value
-    else:
-        obj_out = txt
-    return obj_out