+++ tests/test_printprogress.py
@@ -23,22 +23,22 @@ def test_content(response):
     # assert 'GitHub' in BeautifulSoup(response.content).title.string
 
 def test_printprogress():
-    N = 15000000
+    N = 3000000
     pprog = printprogress(N)
     for _ in range(N):
         pprog()
     # assert input('Did it show you a progress bar? (y for yes)')=='y'
 
-def test_with_logger():
+def test_printprogress_with_logger():
     logger = lognflow()
     N = 1500000
     pprog = printprogress(N, print_function = logger, log_time_stamp = False)
     for _ in range(N):
         pprog()
         
-def test_ETA():
+def test_printprogress_ETA():
     logger = lognflow()
-    N = 1500000
+    N = 500000
     pprog = printprogress(N, print_function = None)
     for _ in range(N):
         ETA = pprog()
@@ -46,7 +46,7 @@ def test_ETA():
     
 
 if __name__ == '__main__':
-    test_ETA()
+    test_printprogress_with_logger()
+    test_printprogress_ETA()
     test_printprogress()
-    test_with_logger()
     exit()